修改列表样式

+++ src/app/manage/knowledge/knowledge.component.ts
@@ -4,37 +4,44 @@ import { ApiService } from 'src/app/api.service';
 import { Knowledge } from 'src/app/entities/knowledge';
 import { BaseResult, KnowledgeResult } from 'src/app/entities/Result';
 import { Observable } from 'rxjs';
+import { ManageService } from 'src/app/shared/manage.service';
 
 declare let NativeObj
 
 @Component({
   selector: 'app-knowledge',
   templateUrl: './knowledge.component.html',
-  styleUrls: ['./knowledge.component.scss']
+  styleUrls: ['../manage.scss', './knowledge.component.scss']
 })
 export class KnowledgeComponent implements OnInit {
 
   constructor(
     private route: ActivatedRoute,
-    private api: ApiService
+    private api: ApiService,
+    private manageService: ManageService
   ) { }
 
+  knowledgeAmount: number
   page: number
   knowledges: Knowledge[] = []
-  counter: number[]
   errText: string
   deletingKnowledge: Knowledge
+//每页数量
+  readonly itemCount = this.manageService.itemCount
+  pageCount
 
   ngOnInit() {
-    this.page = Number(this.route.snapshot.paramMap.get('page'))
-    this.counter = this.api.makeCounterArray(NativeObj.GetKnowledgeCount() / 5)
-    if (!this.page)
-      this.page = 1
+    this.page = Number(this.route.snapshot.queryParamMap.get('page'))
+    if (!this.page) { this.page = 1 }
+    this.api.getStatisticData().toPromise().then(data => {
+      this.knowledgeAmount = (data && data.knowledge) ? data.knowledge.number : 0
+      this.pageCount = Math.ceil(this.knowledgeAmount / this.itemCount)
+    })
     this.loadKnowledges(this.page)
   }
 
   loadKnowledges(page: number) {
-    this.api.getKnowledges(page, 5).subscribe((res: KnowledgeResult) => {
+    this.api.getKnowledges(page, this.itemCount > 0 ? this.itemCount : 1).subscribe((res: KnowledgeResult) => {
       if (res.error) {
         this.errText = res.error
         return
@@ -58,7 +65,8 @@ export class KnowledgeComponent implements OnInit {
         knowledge.isFreezed = !knowledge.isFreezed
     })
   }
-  deleteKnowledge() {
+  deleteKnowledge(knowledge: Knowledge) {
+    if (!confirm(`删除地质科普${knowledge.name}？注意，此操作不可撤销。`)) { return }
     let obs = this.api.deleteKnowledge(this.deletingKnowledge.code)
     obs.subscribe((res: BaseResult) => {
       if (res.error)
@@ -68,7 +76,9 @@ export class KnowledgeComponent implements OnInit {
       this.deletingKnowledge = null
     })
   }
-  setDeletingKnowledge(knowledge: Knowledge) {
-    this.deletingKnowledge = knowledge
+  
+  //页面跳转
+  jump(index: number) {
+    this.loadKnowledges(index)
   }
 }
